readPackage and readPackageScope
