+++ lib/internal/modules/cjs/loader.js
@@ -55,7 +55,6 @@ const {
   StringPrototypeCharAt,
   StringPrototypeCharCodeAt,
   StringPrototypeEndsWith,
-  StringPrototypeLastIndexOf,
   StringPrototypeIndexOf,
   StringPrototypeRepeat,
   StringPrototypeSlice,
@@ -68,7 +67,7 @@ const cjsParseCache = new SafeWeakMap();
 
 // Set first due to cycle with ESM loader functions.
 module.exports = {
-  wrapSafe, Module, readPackageScope, cjsParseCache,
+  wrapSafe, Module, cjsParseCache,
   get hasLoadedAnyUserCJSModule() { return hasLoadedAnyUserCJSModule; },
   initializeCJS,
 };
@@ -89,7 +88,6 @@ const { internalCompileFunction } = require('internal/vm');
 const assert = require('internal/assert');
 const fs = require('fs');
 const path = require('path');
-const { sep } = path;
 const { internalModuleStat } = internalBinding('fs');
 const { safeGetenv } = internalBinding('credentials');
 const {
@@ -403,15 +401,7 @@ function initializeCJS() {
 //   -> a.<ext>
 //   -> a/index.<ext>
 
-/**
- * @param {string} requestPath
- * @return {PackageConfig}
- */
-function readPackage(requestPath) {
-  return packageJsonReader.read(path.resolve(requestPath, 'package.json'));
-}
-
-let _readPackage = readPackage;
+let _readPackage = packageJsonReader.readPackage;
 ObjectDefineProperty(Module, '_readPackage', {
   __proto__: null,
   get() { return _readPackage; },
@@ -422,37 +412,6 @@ ObjectDefineProperty(Module, '_readPackage', {
   },
   configurable: true,
 });
-
-/**
- * Get the nearest parent package.json file from a given path.
- * Return the package.json data and the path to the package.json file, or false.
- * @param {string} checkPath The path to start searching from.
- */
-function readPackageScope(checkPath) {
-  const rootSeparatorIndex = StringPrototypeIndexOf(checkPath, sep);
-  let separatorIndex;
-  const enabledPermission = permission.isEnabled();
-  do {
-    separatorIndex = StringPrototypeLastIndexOf(checkPath, sep);
-    checkPath = StringPrototypeSlice(checkPath, 0, separatorIndex);
-    // Stop the search when the process doesn't have permissions
-    // to walk upwards
-    if (enabledPermission && !permission.has('fs.read', checkPath + sep)) {
-      return false;
-    }
-    if (StringPrototypeEndsWith(checkPath, sep + 'node_modules')) {
-      return false;
-    }
-    const pjson = _readPackage(checkPath + sep);
-    if (pjson.exists) {
-      return {
-        data: pjson,
-        path: checkPath,
-      };
-    }
-  } while (separatorIndex > rootSeparatorIndex);
-  return false;
-}
 
 /**
  * Try to load a specifier as a package.
@@ -574,7 +533,7 @@ function trySelfParentPath(parent) {
 function trySelf(parentPath, request) {
   if (!parentPath) { return false; }
 
-  const { data: pkg, path: pkgPath } = readPackageScope(parentPath);
+  const { data: pkg, path: pkgPath } = packageJsonReader.readPackageScope(parentPath);
   if (!pkg || pkg.exports == null || pkg.name === undefined) {
     return false;
   }
@@ -1134,7 +1093,7 @@ Module._resolveFilename = function(request, parent, isMain, options) {
 
   if (request[0] === '#' && (parent?.filename || parent?.id === '<repl>')) {
     const parentPath = parent?.filename ?? process.cwd() + path.sep;
-    const pkg = readPackageScope(parentPath) || { __proto__: null };
+    const pkg = packageJsonReader.readPackageScope(parentPath) || { __proto__: null };
     if (pkg.data?.imports != null) {
       try {
         const { packageImportsResolve } = require('internal/modules/esm/resolve');
@@ -1431,7 +1390,7 @@ Module._extensions['.js'] = function(module, filename) {
     content = fs.readFileSync(filename, 'utf8');
   }
   if (StringPrototypeEndsWith(filename, '.js')) {
-    const pkg = readPackageScope(filename) || { __proto__: null };
+    const pkg = packageJsonReader.readPackageScope(filename) || { __proto__: null };
     // Function require shouldn't be used in ES modules.
     if (pkg.data?.type === 'module') {
       const parent = moduleParentCache.get(module);
