+++ lib/internal/modules/package_json_reader.js
@@ -4,12 +4,17 @@ const {
   JSONParse,
   ObjectPrototypeHasOwnProperty,
   SafeMap,
+  StringPrototypeEndsWith,
+  StringPrototypeIndexOf,
+  StringPrototypeLastIndexOf,
+  StringPrototypeSlice,
 } = primordials;
 const {
   ERR_INVALID_PACKAGE_CONFIG,
 } = require('internal/errors').codes;
 const { internalModuleReadJSON } = internalBinding('fs');
-const { toNamespacedPath } = require('path');
+const { resolve, sep, toNamespacedPath } = require('path');
+const permission = require('internal/process/permission');
 const { kEmptyObject, setOwnProperty } = require('internal/util');
 
 const { fileURLToPath, pathToFileURL } = require('internal/url');
@@ -111,4 +116,47 @@ function read(jsonPath, { base, specifier, isESM } = kEmptyObject) {
   return result;
 }
 
-module.exports = { read };
+/**
+ * @param {string} requestPath
+ * @return {PackageConfig}
+ */
+function readPackage(requestPath) {
+  return read(resolve(requestPath, 'package.json'));
+}
+
+/**
+ * Get the nearest parent package.json file from a given path.
+ * Return the package.json data and the path to the package.json file, or false.
+ * @param {string} checkPath The path to start searching from.
+ */
+function readPackageScope(checkPath) {
+  const rootSeparatorIndex = StringPrototypeIndexOf(checkPath, sep);
+  let separatorIndex;
+  const enabledPermission = permission.isEnabled();
+  do {
+    separatorIndex = StringPrototypeLastIndexOf(checkPath, sep);
+    checkPath = StringPrototypeSlice(checkPath, 0, separatorIndex);
+    // Stop the search when the process doesn't have permissions
+    // to walk upwards
+    if (enabledPermission && !permission.has('fs.read', checkPath + sep)) {
+      return false;
+    }
+    if (StringPrototypeEndsWith(checkPath, sep + 'node_modules')) {
+      return false;
+    }
+    const pjson = readPackage(checkPath + sep);
+    if (pjson.exists) {
+      return {
+        data: pjson,
+        path: checkPath,
+      };
+    }
+  } while (separatorIndex > rootSeparatorIndex);
+  return false;
+}
+
+module.exports = {
+  read,
+  readPackage,
+  readPackageScope,
+};
